+++ app/src/scenes/Soliguide/Soliguide.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import WebView from 'react-native-webview';
+
+const Soliguide = () => {
+  return <WebView source={{ uri: 'https://soliguide.fr/' }} />;
+};
+
+export default Soliguide;